requirements.txt
fastapi
uvicorn
requests
pandas
ultralytics
Pillow